put quick reply modal content into quick reply modal HOC
